Part 9.1

+++ part-4/4.1/controllers/blogs.js
@@ -1,15 +1,5 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
-const User = require('../models/user')
-const jwt = require('jsonwebtoken')
-
-const getTokenFrom = request => {
-    const authorization = request.get('authorization')
-    if (authorization && authorization.startsWith('Bearer ')) {
-        return authorization.replace('Bearer ', '')
-    }
-    return null
-}
 
 blogsRouter.get('/', async (request, response) => {
     const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
@@ -19,17 +9,9 @@ blogsRouter.get('/', async (request, response) => {
 
 blogsRouter.post('/', async (request, response) => {
     const body = request.body
+    const user = request.user
 
     try {
-        const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
-        if (!decodedToken.id) {
-            return response.status(401).json({ error: 'token missing or invalid' })
-        }
-
-        const user = await User.findById(decodedToken.id)
-        if (!user) {
-            return response.status(401).json({ error: 'user not found' })
-        }
 
         user.blogs = user.blogs || []
 
@@ -47,7 +29,6 @@ blogsRouter.post('/', async (request, response) => {
 
         const savedBlog = await blog.save()
         // console.log('savedBlog', savedBlog._id.toString())
-        user.blogs = user.blogs.concat(savedBlog._id.toString())
 
         user.blogs = user.blogs || []
         user.blogs = user.blogs.concat(savedBlog._id)
@@ -55,19 +36,36 @@ blogsRouter.post('/', async (request, response) => {
 
         response.status(201).json(savedBlog)
     } catch (error) {
-        console.error(error)
+        // console.error(error)
         response.status(400).json({ error: 'bad request' })
     }
 })
 
-blogsRouter.delete('/:id', (request, response) => {
-    Blog.findByIdAndRemove(request.params.id)
-        .then(() => {
-            response.status(204).end()
-        })
-        .catch(() => {
-            response.status(400).json('Bad request')
-        })
+blogsRouter.delete('/:id', async (request, response) => {
+
+    // console.log(request.user)
+
+    const user = request.user
+
+    const blog = await Blog.findById(request.params.id)
+
+    if (!blog) {
+        return response.status(401).json({ error: 'Blog not found' })
+    }
+
+    if (blog.user.toString() === user.id.toString()) {
+        Blog.findByIdAndRemove(request.params.id)
+            .then(() => {
+                response.status(204).end()
+            })
+            .catch(() => {
+                response.status(400).json('Bad request')
+            })
+        // return response.status(401).json({ error: 'azerhjksfghb' })
+    } else {
+        return response.status(401).json({ error: 'user not authorized' })
+    }
+
 })
 
 blogsRouter.put('/:id', async (request, response) => {
